+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.28.2] - 2020-12-03
+
 ### Changed
 
 - Use trade policy from selected facets in `productSearch` query.